requirements.txt
# requirements.txt
azure-identity>=1.15.0
msgraph-sdk>=1.0.0
pandas>=2.0.0
colorama>=0.4.6
aiohttp>=3.9.0
python-dotenv>=1.0.0
pillow>=10.0.0